# backend-broker/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BitCurrent API",
    "description": "UK Crypto Broker API - Trade Bitcoin, Ethereum & more with 0.25% fees",
    "version": "1.0.0",
    "contact": {
      "email": "support@bitcurrent.co.uk",
      "url": "https://bitcurrent.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://bitcurrent-production.up.railway.app/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "post": {
        "tags": ["Trading"],
        "summary": "Place new order",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string", "example": "BTC-GBP" },
                  "side": { "type": "string", "enum": ["buy", "sell"] },
                  "quantity": { "type": "number", "example": 0.01 },
                  "type": { "type": "string", "enum": ["market"], "default": "market" }
                },
                "required": ["symbol", "side", "quantity"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed successfully"
          },
          "400": {
            "description": "Invalid order parameters"
          },
          "403": {
            "description": "Insufficient funds"
          }
        }
      },
      "get": {
        "tags": ["Trading"],
        "summary": "Get user's orders",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of user's orders"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}